+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.4] - 2026-01-19
+
+### Added
+- Custom SVG icon support for Bluetooth devices
+- Smart icon detection: AirPods use SF Symbol, other Bluetooth devices use custom bluetooth_speaker.svg
+- Asset catalog with bluetooth_speaker icon
+
+### Changed
+- Bluetooth devices now show appropriate icons based on device name
+
 ## [1.0.3] - 2026-01-19
 
 ### Added